Nexus: Finalize removal of old `object_interface` and `obj`
